Rename custom trustmanager test package name
Previous package name used camel case. Corrects naming to be lower case.
