+++ src/test/java/com/microsoft/sqlserver/jdbc/ssl/trustmanager/PermissiveTrustManager.java
@@ -1,14 +1,16 @@
-package com.microsoft.sqlserver.jdbc.CustomTrustManager;
+package com.microsoft.sqlserver.jdbc.ssl.trustmanager;
 
 import java.security.cert.CertificateException;
 import java.security.cert.X509Certificate;
 
+import javax.net.ssl.X509TrustManager;
+
 /**
- * This class does not implement X509TrustManager and the connection must fail when it is specified by the trustManagerClass property
- * 
+ * This class implements an X509TrustManager that always accepts the X509Certificate chain offered to it.
  */
-public final class InvalidTrustManager {
-    public InvalidTrustManager() {
+
+public final class PermissiveTrustManager extends Object implements X509TrustManager {
+    public PermissiveTrustManager() {
     }
 
     public void checkClientTrusted(X509Certificate[] chain,
