+++ src/test/java/com/microsoft/sqlserver/jdbc/ssl/trustmanager/TrustManagerWithConstructorArg.java
@@ -1,4 +1,4 @@
-package com.microsoft.sqlserver.jdbc.CustomTrustManager;
+package com.microsoft.sqlserver.jdbc.ssl.trustmanager;
 
 import java.io.IOException;
